refactor: 컨트롤러를 인터페이스와 구현체로 분리
